refactor: set srcDir
