+++ package.json
@@ -31,9 +31,11 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "date-fns": "^3.6.0",
+    "idb-keyval": "^6.2.1",
     "js-md5": "^0.8.3",
     "ofetch": "^1.3.4",
-    "tailwind-merge": "^2.2.2"
+    "tailwind-merge": "^2.2.2",
+    "unstorage": "^1.10.2"
   },
   "devDependencies": {
     "@antfu/eslint-config": "2.14.0",
