+++ src/components/BaseStatus.vue
