+++ src/components/CharacterSwitcher.vue
