+++ src/composables/animations/animate-cursor.ts
