+++ src/entrypoints/popup/Popup.vue
