+++ src/utils/background/storage.ts
@@ -1,0 +1,6 @@
+import { createStorage } from 'unstorage'
+import indexedDbDriver from 'unstorage/drivers/indexedb'
+
+export const indexedDbStorage = createStorage({
+  driver: indexedDbDriver({ base: 'app:' }),
+})
