+++ unocss.config.ts
@@ -9,7 +9,7 @@ export default defineConfig<Theme>({
         // the default
         /\.(vue|svelte|[jt]sx|mdx?|astro|elm|php|phtml|html)($|\?)/,
         // include js/ts files
-        'components/**/*.ts',
+        'src/components/**/*.ts',
       ],
     },
   },
